Cosmetics commit_hash:6706f266797a03238924fea9d1b890e244542b0f

+++ yt/yt/server/tablet_balancer/config.h
@@ -26,7 +26,7 @@ public:
 
     TDuration Period;
 
-    i64 WorkerThreadPoolSize;
+    int WorkerThreadPoolSize;
     i64 PivotPickerThreadPoolSize;
 
     TDuration ParameterizedTimeoutOnStart;
